Fix: Add, Edit, Delete on search data issue
